+++ server/src/test/java/com/inventage/portal/gateway/proxy/middleware/bearerOnly/BearerOnlyMiddlewareOtherClaimsTest.java
@@ -8,6 +8,7 @@ import com.google.common.io.Resources;
 import com.inventage.portal.gateway.proxy.middleware.authorization.bearerOnly.customClaimsChecker.JWTAuthAdditionalClaimsOptions;
 import com.inventage.portal.gateway.proxy.middleware.authorization.bearerOnly.customClaimsChecker.JWTClaim;
 import com.inventage.portal.gateway.proxy.middleware.authorization.bearerOnly.customClaimsChecker.JWTClaimOperator;
+import com.inventage.portal.gateway.proxy.middleware.authorization.bearerOnly.customIssuerChecker.JWTAuthProviderIssuer;
 import com.inventage.portal.gateway.proxy.middleware.mock.TestBearerOnlyJWTProvider;
 import io.vertx.core.Vertx;
 import io.vertx.core.http.HttpHeaders;
@@ -233,7 +234,7 @@ public class BearerOnlyMiddlewareOtherClaimsTest {
         } catch (IOException e) {
             e.printStackTrace();
         }
-        return JWTAuth.create(vertx, new JWTAuthOptions()
+        return JWTAuthProviderIssuer.create(vertx, new JWTAuthOptions()
             .addPubSecKey(new PubSecKeyOptions().setAlgorithm(publicKeyAlgorithm)
                 .setBuffer(publicKeyRS256))
             .setJWTOptions(new JWTOptions()